progression: use long level aggregation

+++ app/src/main/java/com/zerotoempire/game/ProgressionSystems.kt
@@ -67,7 +67,7 @@ object Progression {
     fun achievements(state: GameState, meta: PlayerMeta) = listOf(
         Achievement("first", "First Step", "Own your first business", state.businesses.any { it.level > 0 }, "first" in meta.claimedAchievementIds, 5),
         Achievement("million", "Millionaire", "Earn 1M lifetime cash", state.lifetimeCash >= 1_000_000.0, "million" in meta.claimedAchievementIds, 10),
-        Achievement("century", "Industrial Machine", "Own 100 total business levels", state.businesses.sumOf { it.level } >= 100, "century" in meta.claimedAchievementIds, 15),
+        Achievement("century", "Industrial Machine", "Own 100 total business levels", state.businesses.sumOf { it.level.toLong() } >= 100L, "century" in meta.claimedAchievementIds, 15),
         Achievement("reborn", "Reborn", "Prestige for the first time", meta.prestigeCount > 0, "reborn" in meta.claimedAchievementIds, 20)
     )
 }
